Merge upstream!

+++ src/main/java/io/inscopemetrics/kairosdb/HistogramDataPoint.java
@@ -29,13 +29,13 @@ public interface HistogramDataPoint extends DataPoint {
      * Getter the number of data points originally contained in this HistogramDataPoint.
      * @return datapoint sample count
      */
-    int getOriginalCount();
+    long getOriginalCount();
 
     /**
      * Getter for sample count contained in this HistogramDataPoint.
      * @return datapoint sample count
      */
-    int getSampleCount();
+    long getSampleCount();
 
     /**
      * Getter for the sum of all samples contained in this HistogramDataPoint.
